server/requirements.txt
Flask
flask-cors
openai
openai-whisper
python-dotenv
yt-dlp
certifi
